chore(kuma-cp) generify proxy template matching
